requirements.txt
torch
pillow
numpy
open-cv